+++ docs/CHANGELOG.md
@@ -1,28 +1,37 @@
 #### **Version 0.1.7:**
 
+
 * Fixed some bugs
 * Cleared code in PyCMD.py
 
+
 #### **Version 0.1.8:**
+
 
 * Added system required to README.md
 * Added CHANGELOG.md
 * Moved some modules to folder "modules"
 * Moved CHANGELOG.md and README.md to folder "docs"
 
+
 #### **Version 0.1.9:**
+
 
 * Fixed some bugs
 * Added command "config"
 * Updated help screen
 
+
 #### **Version 0.2:**
+
 
 * Fixed some bugs
 * Added mods system
 * Added logs system
 
+
 #### **Version 1.0:**
+
 
 * Official release
 * Fixed some bugs
@@ -30,7 +39,9 @@
 * Plans for next ver: 
     * Add system for turning off mods 
 
+
 #### **Version 1.1.1:**
+
 
 * Done plans:
     * Fixed some bugs
@@ -39,62 +50,98 @@
 * Plans for v1.2:
     * Add updater
 
+
 #### **Version 1.2:**
+
 
 * Done plans:
     * Fixed some bugs
     * Added system for updating from official repository
 * Added prefix for command prompt
 
+
 #### **Version 1.2.1:**
+
 
 * Fixed critical bug
 
+
 #### **Version 1.3:**
+
 
 * Added command for downloading mods
 * Updated help screen
 
+
 #### **Version 1.4:**
+
 
 * Added command for reset PyCMD
 * Updated help screen
 * Fixed some bugs in PyCMD and getmods
 
+
 #### **Version 1.4.1:**
+
 
 * Added support for only startup mods
 * Fixed some bugs
 
+
 #### **Version 1.4.2**
+
 
 * Change updater location to main folder, so now you can start it from file
 * From now updater is not deleting your mods
 
+
 #### **Version 2.0:**
+
 
 * Added fabric settings
 * Added admin system
 * Changed way which You use to install mods
 * Fixed some bugs
 
+
 #### **Version 2.0.1:**
+
 
 * Fixed security bug
 
+
 #### **Version 2.0.2:**
+
 
 * Fixed critical bug
 
+
 #### **Version 2.0.3:**
+
 
 * Fixed fabric bugs
 * Added way to log out from admin account
 
+
 #### **Version 2.0.4:**
+
 
 * Fixed first app bugs
 
+
 #### **Version 2.0.5:**
 
+
 * Fixed style for admin mode
+
+
+#### **Version 2.1:**
+
+
+* Updated help screen
+* Added cmd command
+* Added changepass system
+* Added changeprefix system
+
+* PLANS:
+    * Repair mods uninstall system